Refactor the Guardian class
break out different parts of functionality into different files and make the Guardian class partial in order to segregate different pieces of functionality. The class is cumbersome and difficult to navigate and so we break it out into constituent pieces so that it is easier to understand what is happening. In the future we may want to consider refactoring this object entirely so that it does not violate the single responsibility principle.
